CDAP-11586 update start script to handle spark2
The assumption is that the environment will contain
SPARK_MAJOR_VERSION=2 when CDAP should be using spark2. This is
the same as how HDP does it.

Modified the script to set SPARK_COMPAT, and to set java system
property app.program.spark.compat, which CDAP uses to decide
which spark runtime to use, and which pipeline artifacts to
load.